+++ src/api/domain/idCard.api.server.ts
@@ -1,5 +1,8 @@
 import { privateApi } from '~/api/config/privateApi.server';
-import { IdCardDetailResponse } from '~/types/idCard';
+import { CommunityMyIdCardDetailResponse, IdCardDetailResponse } from '~/types/idCard';
 
 export const getIdCardDetailServer = (idCardId: number) =>
   privateApi.get<IdCardDetailResponse>(`/id-cards/${idCardId}`);
+
+export const getCommunityMyIdCardDetailServer = (communityId: number) =>
+  privateApi.get<CommunityMyIdCardDetailResponse>(`/communities/${communityId}/users/idCards`);
