+++ src/app/my-page/[id]/components/MyPageIdCard/index.ts
@@ -1,0 +1,1 @@
+export * from './MyPageIdCard';
